+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 History
 -------
 
+ - Add parameter noecho to inputs.read
+ - Add functions walign, wlen, wshorten to module strings
+
 **2020-07-30 (0.14.0)**
  - Add parameter maxsplit to strings.str2tuple
  - Deprecate modules crypto and config